+++ CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
-- Added dialog for choosing contact source while editing a merged contact ([#201])
+- Dialog for choosing contact source while editing a merged contact ([#201])
 
 ## [1.1.0] - 2024-10-28